# server/openapi.yaml
openapi: 3.0.3
info:
  title: AIBillingSystem API
  description: |
    API for the AIBillingSystem - a comprehensive medical billing and practice management system
    designed for therapy practices. Features include patient management, claims processing,
    appointment scheduling, analytics, and AI-powered billing optimization.

    ## Authentication
    Most endpoints require authentication via session-based auth (Replit Auth).
    Protected endpoints will return 401 Unauthorized if not authenticated.

    ## Multi-Tenancy
    The system supports multiple practices. Users are assigned to a practice and can only
    access data within their authorized practice (except admin users who have cross-practice access).

    ## HIPAA Compliance
    This API handles Protected Health Information (PHI). All access is logged for audit purposes.
  version: 1.0.0
  contact:
    name: AIBillingSystem Support
  license:
    name: Proprietary
servers:
  - url: /api
    description: API base path
...
paths:
...
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get dashboard analytics
      description: Returns key metrics for the dashboard. Financial data only visible to admin/billing roles.
      operationId: getDashboardAnalytics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dashboard analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardAnalytics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/revenue:
    get:
      tags:
        - Analytics
      summary: Get revenue analytics
      description: Returns revenue data over a specified time range
      operationId: getRevenueAnalytics
      security:
        - sessionAuth: []
      parameters:
        - name: timeRange
          in: query
          schema:
            type: string
            enum: ['3months', '6months', '12months']
            default: '12months'
          description: Time range for the analytics
      responses:
        '200':
          description: Revenue analytics data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    month:
                      type: string
                    revenue:
                      type: number
                    claims:
                      type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/claims-by-status:
    get:
      tags:
        - Analytics
      summary: Get claims by status
      description: Returns claim counts grouped by status
      operationId: getClaimsByStatus
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Claims by status
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    status:
                      type: string
                    count:
                      type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/denial-reasons:
    get:
      tags:
        - Analytics
      summary: Get denial reasons analytics
      description: Returns common claim denial reasons with counts
      operationId: getDenialReasons
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Denial reasons with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    reason:
                      type: string
                    count:
                      type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/collection-rate:
    get:
      tags:
        - Analytics
      summary: Get collection rate
      description: Returns the percentage of billed amounts that were collected
      operationId: getCollectionRate
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Collection rate data
          content:
            application/json:
              schema:
                type: object
                properties:
                  rate:
                    type: number
                    format: decimal
                  totalBilled:
                    type: number
                  totalCollected:
                    type: number
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/clean-claims-rate:
    get:
      tags:
        - Analytics
      summary: Get clean claims rate
      description: Returns the percentage of claims submitted without errors
      operationId: getCleanClaimsRate
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Clean claims rate data
          content:
            application/json:
              schema:
                type: object
                properties:
                  rate:
                    type: number
                    format: decimal
                  cleanClaims:
                    type: integer
                  totalClaims:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/capacity:
    get:
      tags:
        - Analytics
      summary: Get practice capacity analytics
      description: Returns therapist capacity and utilization metrics
      operationId: getCapacityAnalytics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Capacity analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalSlots:
                    type: integer
                  bookedSlots:
                    type: integer
                  utilizationRate:
                    type: number
                    format: decimal
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/ar-aging:
    get:
      tags:
        - Analytics
      summary: Get accounts receivable aging
      description: Returns AR aging buckets (0-30, 31-60, 61-90, 90+ days)
      operationId: getARAgingAnalytics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: AR aging data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    bucket:
                      type: string
                    amount:
                      type: number
                    count:
                      type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/cancellations:
    get:
      tags:
        - Analytics
      summary: Get cancellation analytics
      description: Returns appointment cancellation metrics
      operationId: getCancellationAnalytics
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Cancellation analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalCancellations:
                    type: integer
                  cancellationRate:
                    type: number
                    format: decimal
                  byReason:
                    type: array
                    items:
                      type: object
                      properties:
                        reason:
                          type: string
                        count:
                          type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication via Replit Auth
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Unauthorized
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: An unexpected error occurred
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
...
    DashboardAnalytics:
      type: object
      properties:
        totalPatients:
          type: integer
          description: Total number of active patients
        activeClaims:
          type: integer
          description: Number of claims in progress
        pendingPayments:
          type: integer
          description: Number of claims awaiting payment
        claimApprovalRate:
          type: number
          format: decimal
          description: Percentage of claims approved
        monthlyRevenue:
          type: number
          format: decimal
          description: Revenue for current month (admin/billing only)
        averageReimbursement:
          type: number
          format: decimal
          description: Average reimbursement per claim (admin/billing only)
security:
  - sessionAuth: []